backend/requirements.txt
# Core Flask dependencies
Flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.5
python-socketio==5.10.0
python-engineio==4.8.0
werkzeug==3.0.1

# Image processing
opencv-python==4.10.0.84
numpy==2.1.1
Pillow>=11.0.0

# OCR
pytesseract==0.3.10

# PDF generation and manipulation
fpdf2==2.7.9
PyMuPDF>=1.23.0
reportlab>=4.0.0

# Machine Learning (for document classification)
scikit-learn>=1.3.0

# Cloud storage (optional)
boto3>=1.28.0

# Windows printing
# Windows specific (for printing)
pywin32>=307
